+++ src/main/kotlin/org/lunakoly/quicklink/repository/RepositoryInfo.kt
@@ -7,7 +7,4 @@ class RepositoryInfo(
     @Suppress("unused")
     val branch: String?,
     val commitHash: String,
-    val remotes: Map<String, String>,
-) {
-    val remotesNames get() = remotes.keys.toList()
-}
+)